requirements.txt
# Core dependencies
numpy>=1.24.0
torch>=2.0.0
skia-python>=87.5
Pillow>=10.0.0

# Video processing
opencv-python>=4.8.0
ffmpeg-python>=0.2.0

# Utilities
pydantic>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0
loguru>=0.7.0

# Optional performance enhancements
# cupy-cuda11x>=12.0.0  # Uncomment if using CuPy
# numba>=0.57.0  # Uncomment if using Numba

